implement oauth/jwt auth on server for web
This commit introduces new web.Auth structure which abstracts how we
handle authorization on server side providing 3 public methods:

- Login: entrypoint for unauthorized user
- Callback: auth callback endpoint
- Middleware: allows to protect http.Handler and sets userID into
request

I didn't abstract OAuth provider because currently we are supporting
only github but it must be very easy to do when we need it.
Only constructor and getUser method depend on it.

The code is almost indentical to the code in code-annotation project but
incapsulates everything in one structure as I don't see any reason to
complicate it with more flexibility.

Ref: https://github.com/src-d/lookout/issues/399

Signed-off-by: Maxim Sukharev <max@smacker.ru>