remove TestServerCancel test
It doesn't test our code and makes an incorrect assumption about how grpc works.

The logic is:
- create a response of N items
- create DataServerHandler which implements grpc interface for an iterator
- create grpc Server (vendored code)
- create grpc DataClient client (vendored code)
- calls grpc method on DataClient (vendored code) with cancellable context
- read [0, N-1] items from the response stream
- send cancel
- check that client read returns any error containing "context cancel"

The actual error it receives is "statusError" coming from grpc vendored code.
Context gets propagated to the stream and in the stream constructor
it runs goroutine which finishes the stream on context cancel.
When the stream is finished it just returns the reason error for all calls.

So it doesn't matter what server is used and if it handles cancellation or not.

It also makes an incorrect assumption that after `cancel`
next call is guaranteed to return an error.
As explained above handling of context cancelation happens in goroutine,
so scheduler *can* run `Recv` before `stream.finish`
which we see as flaky test behavior on CI.

I believe the original reason for this test was to test `DataServerHandler` code
and one can actually see test coverage decrease after removing this test *sometimes*.

It happens because of the side effects:
- Server sometimes hits an error while writing into the stream using `Send` method.
  It can be `context canceled` or `transport is closing` on my machine.
- It also hits <-cancel path but not very often

Though the code doesn't check any of them.

Ref: #519

Signed-off-by: Maxim Sukharev <max@smacker.ru>